requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
pypdf
python-dotenv
groq
langchain-community